added bake system update.

+++ lib/bake/cli/system.ex
@@ -19,6 +19,7 @@ defmodule Bake.Cli.System do
 
     case cmd do
       ["get"] -> get(opts)
+      ["update"] -> update(opts)
       ["clean"] -> clean(opts)
       _ -> invalid_cmd(cmd)
     end
@@ -45,13 +46,8 @@ defmodule Bake.Cli.System do
             Bake.Api.System.get(%{recipe: v[:recipe]})
             |> get_resp(platform: platform, adapter: adapter, lock_file: lock_path, target: target)
           [{recipe, version}] ->
-            system_path = "#{adapter.systems_path}/#{recipe}-#{version}"
-            if File.dir?(system_path) do
-              Bake.Shell.info "=> System #{recipe} at #{version} up to date"
-            else
-              Bake.Api.System.get(%{recipe: recipe, version: version})
-              |> get_resp(platform: platform, adapter: adapter, lock_file: lock_path, target: target)
-            end
+            Bake.Api.System.get(%{recipe: recipe, version: version})
+            |> get_resp(platform: platform, adapter: adapter, lock_file: lock_path, target: target)
         end
       else
         # The lockfile doesn't exist. Download latest version
@@ -61,26 +57,47 @@ defmodule Bake.Cli.System do
     end)
   end
 
+  defp update(opts) do
+    {bakefile_path, target_config, target} = bakefile(opts[:bakefile], opts[:target])
+    platform = target_config[:platform]
+    adapter = adapter(platform)
+
+    lock_path = bakefile_path
+    |> Path.dirname
+    lock_path = lock_path <> "/Bakefile.lock"
+
+    Enum.each(target_config[:target], fn {target, v} ->
+      {recipe, version} = v[:recipe]
+      Bake.Api.System.get(%{recipe: recipe, requirement: version})
+      |> get_resp(platform: platform, adapter: adapter, lock_file: lock_path, target: target)
+    end)
+  end
+
   defp get_resp({:ok, %{status_code: code, body: body}}, opts) when code in 200..299 do
     %{data: %{path: path, host: host, name: name, version: version, username: username}} = Poison.decode!(body, keys: :atoms)
 
     adapter = opts[:adapter]
-    lock_file = opts[:lock_file]
-    target = opts[:target]
-    Bake.Shell.info "=> Downloading system #{username}/#{name}-#{version}"
-    case Bake.Api.request(:get, host <> "/" <> path, []) do
-      {:ok, %{status_code: code, body: tar}} when code in 200..299 ->
-        Bake.Shell.info "=> System #{username}/#{name}-#{version} downloaded"
-        dir = adapter.systems_path <> "/#{username}"
-        File.mkdir_p(dir)
-        File.write!("#{dir}/#{name}-#{version}.tar.gz", tar)
-        Bake.Shell.info "=> Unpacking system #{username}/#{name}-#{version}"
-        System.cmd("tar", ["zxf", "#{name}-#{version}.tar.gz"], cd: dir)
-        File.rm!("#{dir}/#{name}-#{version}.tar.gz")
-        Bake.Config.Lock.write(lock_file, [targets: [{target, ["#{username}/#{name}": version]}]])
-      {_, response} ->
-        Bake.Shell.error("Failed to download system")
-        Bake.Utils.print_response_result(response)
+    system_path = "#{adapter.systems_path}/#{username}/#{name}-#{version}"
+    if File.dir?(system_path) do
+      Bake.Shell.info "=> System #{username}/#{name} at #{version} up to date"
+    else
+      lock_file = opts[:lock_file]
+      target = opts[:target]
+      Bake.Shell.info "=> Downloading system #{username}/#{name}-#{version}"
+      case Bake.Api.request(:get, host <> "/" <> path, []) do
+        {:ok, %{status_code: code, body: tar}} when code in 200..299 ->
+          Bake.Shell.info "=> System #{username}/#{name}-#{version} downloaded"
+          dir = adapter.systems_path <> "/#{username}"
+          File.mkdir_p(dir)
+          File.write!("#{dir}/#{name}-#{version}.tar.gz", tar)
+          Bake.Shell.info "=> Unpacking system #{username}/#{name}-#{version}"
+          System.cmd("tar", ["zxf", "#{name}-#{version}.tar.gz"], cd: dir)
+          File.rm!("#{dir}/#{name}-#{version}.tar.gz")
+          Bake.Config.Lock.write(lock_file, [targets: [{target, ["#{username}/#{name}": version]}]])
+        {_, response} ->
+          Bake.Shell.error("Failed to download system")
+          Bake.Utils.print_response_result(response)
+      end
     end
   end
